ALL GREEN TESTS after refactoring services to be in separate files plus paired mocks - reducer tests done
