+++ src/actions/journal.js
@@ -1,5 +1,5 @@
 import { createAction } from 'promise-middleware-redux';
-import { deleteRequest, getUserNotes } from '../services/herokuApi';
+import { deleteRequest, getUserNotes } from '../services/journal';
 
 export const [
   deleteNote,
@@ -16,9 +16,3 @@ export const [
   FETCH_JOURNAL_LIST_FULFILLED,
   FETCH_JOURNAL_LIST_REJECTED
 ] = createAction('FETCH_JOURNAL_LIST', getUserNotes);
-
-export const UPDATE_TOGGLE = 'UPDATE_TOGGLE';
-export const updateToggle = (toggle) => ({
-  type: UPDATE_TOGGLE,
-  payload: toggle
-});
