+++ src/reducers/message.js
@@ -1,4 +1,4 @@
-import { CREATE_MESSAGE, UPDATE_DATE, UPDATE_MESSAGE, UPDATE_TIME, UPDATE_DAILY, UPDATE_IS_REPEATED, UPDATE_WEEKLY, UPDATE_PRIVATE  } from '../actions/message';
+import { CREATE_MESSAGE, UPDATE_DATE, UPDATE_MESSAGE, UPDATE_TIME, UPDATE_DAILY, UPDATE_IS_REPEATED, UPDATE_WEEKLY, UPDATE_PRIVATE  } from '../actions/message.js';
 
 const initialState = {
   date: '',
