+++ src/selectors/journal.js
@@ -1,3 +1,2 @@
 export const getJournalList = state => state.journal.journalList;
 export const getFavorites = state => state.journal.favorites;
-export const isToggle = state => state.journal.toggle;
