+++ src/services/__mocks__/community.js
@@ -1,0 +1,1 @@
+export const getPublicNotes = () => Promise.resolve(['banana', 'orange', 'apple']);
