+++ src/services/community.js
@@ -1,3 +1,17 @@
-import { get } from './request';
+import store from '../store';
+import { getToken } from '../selectors/session';
 
-export const getCommunityNotes = () => get('/notes');
+export const getPublicNotes = () => {
+  return fetch(`${process.env.API_URL}/notes`, {
+    method: 'GET',
+    headers : {
+      Authorization: `Bearer ${getToken(store.getState())}`,
+      'Content-Type': 'application/json',
+    }
+  })
+    .then(res => [res.ok, res.json()])
+    .then(([ok, json]) => {
+      if(!ok) throw json;
+      return json;
+    });
+};
